[WIP] Work continues on resolving issues with the 2-channel synthesizer that is currently up/down converting the signal of interest, but the synthesizer is not correctly combining the original two channels.
Some experimentation with the tap count for the channelizer indicates
that we can cut the tap count in half and still have decent decoding
on the P25 decoder when paired to an aligned channelized output.  The
2-channel synthesized output decodes partially, but needs more work.

Resolved issue with channel gain in the channelizer that was making the
sample values so small that they were registering as zero when converted
from float to 16-bit int value in the complex wave recorder.